# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Academic Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/ai-tiers:
    get:
      operationId: getAdminAITiers
      tags: [admin]
      summary: "Owner: list all AI tiers including inactive"
      responses:
        "200":
          description: All tiers
          content:
            application/json:
              schema:
                type: object
                properties:
                  tiers:
                    type: array
                    items:
                      $ref: "#/components/schemas/AITierAdmin"
        "401":
          description: Unauthorized
        "403":
          description: Owner access required
  /admin/ai-tiers/{tierId}:
    put:
      operationId: updateAdminAITier
      tags: [admin]
      summary: "Owner: update tier pricing, rate limits, and visibility"
      parameters:
        - name: tierId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAITierRequest"
      responses:
        "200":
          description: Tier updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  tier:
                    $ref: "#/components/schemas/AITierAdmin"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "403":
          description: Owner access required
        "404":
          description: Tier not found

  # ── User Profile ─────────────────────────────────────────────────────────
...
  /account/references:
    get:
      operationId: listAccountReferences
      tags: [account-references]
      summary: List account-level references
      description: Returns all references in the authenticated user's personal reference pool.
      responses:
        "200":
          description: Account references
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountReference"
        "401":
          description: Unauthorized
    post:
      operationId: createAccountReference
      tags: [account-references]
      summary: Add reference to personal pool
      description: Adds a new reference to the user's account-level library. Duplicate DOIs are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountReferenceInput"
      responses:
        "201":
          description: Created reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountReference"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "409":
          description: Reference with this DOI already exists
  /account/references/import:
    post:
      operationId: importAccountReferences
      tags: [account-references]
      summary: Bulk import references from DOIs
      description: Fetches metadata from CrossRef for each DOI and adds valid references to the account library. Skips duplicates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportAccountReferencesRequest"
      responses:
        "201":
          description: Import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportAccountReferencesResponse"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /account/references/{id}:
    put:
      operationId: updateAccountReference
      tags: [account-references]
      summary: Update an account reference
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountReferenceInput"
      responses:
        "200":
          description: Updated reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountReference"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Reference not found
        "409":
          description: Reference with this DOI already exists
    delete:
      operationId: deleteAccountReference
      tags: [account-references]
      summary: Delete an account reference
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Reference not found
  /account/references/{id}/assign:
    post:
      operationId: assignAccountReference
      tags: [account-references]
      summary: Assign account reference to a project
      description: Copies an account-level reference into a project's reference pool.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId]
              properties:
                projectId:
                  type: number
      responses:
        "201":
          description: Reference added to project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Account reference not found
        "409":
          description: Reference already exists in the project
components:
  schemas:
...
    Reference:
      type: object
      required: [id, projectId, title, validationStatus, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        title:
          type: string
        authors:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        journal:
          type: ["string", "null"]
        volume:
          type: ["string", "null"]
        issue:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        validationStatus:
          type: string
          enum: [unverified, verified, invalid]
        usedInChapters:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isSuggested:
          type: boolean
          description: Whether this reference was auto-suggested by CrossRef
        source:
          type: string
          enum: [manual, crossref, file]
          description: Source of the reference
...
    AITierAdmin:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        model:
          type: string
        pricePer1MInputCents:
          type: number
        pricePer1MOutputCents:
          type: number
        providerCostPer1MInputCents:
          type: number
        providerCostPer1MOutputCents:
          type: number
        rateLimitRpm:
          type: ["number", "null"]
        rateLimitTpd:
          type: ["number", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        displayOrder:
          type: number
        description:
          type: string
        usageTips:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateAITierRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        pricePer1MInputCents:
          type: number
          minimum: 0
        pricePer1MOutputCents:
          type: number
          minimum: 0
        providerCostPer1MInputCents:
          type: number
          minimum: 0
        providerCostPer1MOutputCents:
          type: number
          minimum: 0
        rateLimitRpm:
          type: ["number", "null"]
          minimum: 1
          maximum: 10000
        rateLimitTpd:
          type: ["number", "null"]
          minimum: 1
        isFree:
          type: boolean
        isActive:
          type: boolean
        description:
          type: string
          maxLength: 500
        usageTips:
          type: ["string", "null"]
          maxLength: 500
...
    AccountReference:
      type: object
      required: [id, userId, title, createdAt]
      properties:
        id:
          type: number
        userId:
          type: string
        title:
          type: string
        authors:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        journal:
          type: ["string", "null"]
        volume:
          type: ["string", "null"]
        issue:
          type: ["string", "null"]
        doi:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isSuggested:
          type: boolean
          description: Whether this reference was auto-suggested by CrossRef
        source:
          type: string
          enum: [manual, crossref, file]
          description: Source of the reference
    AccountReferenceInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        authors:
          type: string
        year:
          type: number
        journal:
          type: string
        volume:
          type: string
        issue:
          type: string
        doi:
          type: string
        url:
          type: string
        isSuggested:
          type: boolean
          default: false
          description: Whether this reference was auto-suggested by CrossRef
        source:
          type: string
          enum: [manual, crossref, file]
          default: manual
          description: Source of the reference
    ImportAccountReferencesRequest:
      type: object
      required: [dois]
      properties:
        dois:
          type: array
          description: Array of DOI strings to import (max 50)
          items:
            type: string
          minItems: 1
          maxItems: 50
    ImportAccountReferencesResponse:
      type: object
      required: [imported, summary]
      properties:
        imported:
          type: array
          items:
            $ref: "#/components/schemas/AccountReference"
        skipped:
          type: array
          description: DOIs that were already in the user's library
          items:
            type: string
        failed:
          type: array
          description: DOIs that failed to import
          items:
            type: object
            properties:
              doi:
                type: string
              error:
                type: string
        summary:
          type: object
          properties:
            total:
              type: number
            imported:
              type: number
            skipped:
              type: number
            failed:
              type: number